update: move to new craft version
