+++ components/ctas/cta-one.tsx
@@ -8,7 +8,7 @@ import Balancer from "react-wrap-balancer";
 import { Button } from "@/components/ui/button";
 
 // Custom components
-import { Container, Section } from "@/components/craft";
+import { Container, Section } from "@/components/ds";
 
 const CTA = () => {
   return (
