+++ components/faqs/faqs-one.tsx
@@ -1,4 +1,4 @@
-import { Section, Container } from "@/components/craft";
+import { Section, Container } from "@/components/ds";
 
 import { ArrowUpRight } from "lucide-react";
