+++ components/features/feature-eight.tsx
@@ -3,7 +3,7 @@ import Link from "next/link";
 import Balancer from "react-wrap-balancer";
 
 import { Button } from "@/components/ui/button";
-import { Section, Container } from "@/components/craft";
+import { Section, Container } from "@/components/ds";
 import Placeholder from "@/public/placeholder.jpg";
 
 const Feature = () => {
