+++ components/features/feature-nine.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import Image from "next/image";
 
-import { Section, Container } from "@/components/craft";
+import { Section, Container } from "@/components/ds";
 import { Card, CardContent } from "@/components/ui/card";
 
 import {
