+++ components/nav-links.tsx
@@ -1,4 +1,4 @@
-import { Container } from "./craft";
+import { Container } from "./ds";
 import { Button } from "./ui/button";
 import Link from "next/link";
